Barra de vida funcional
Eu fiz alguns comentários no código ajustei umas coisas aqui e ali mas nd de mais
* Arrumar Colisão do player com o cenário. o jogador consegue "entrar" dentro dos blocos sólidos e para de pular

+++ SacraelsArtifacts/Assets/Scripts/UIScripts/HUD/LifeBar.cs
@@ -11,7 +11,7 @@ public class LifeBar : MonoBehaviour
     {
         if (playerHealth != null)
         {
-          LifeBarIMG.fillAmount = playerHealth.health / playerHealth._healthMax;
+          LifeBarIMG.fillAmount = (float)playerHealth.health / playerHealth._healthMax;
         }
     }
 }
